refactor(viewer): divide getting current Measurement logic
